refactor: comment API 분리
